server/requirements.txt
openai-whisper==20231117
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
numpy==1.24.3
ffmpeg-python==0.2.0
